@@ -32,7 +32,8 @@ limitations under the License.
           }
 
           var query = $(this).val();
-          document.location = "{{ "search/" | absURL }}?q=" + query;
+          var searchPage = $(this).data('search-page') + "?q=" + query;
+          document.location = searchPage;
 
           return false;
         });
